исправление

+++ src/main/java/com/example/zvon73/service/BellService.java
@@ -5,6 +5,8 @@ import com.example.zvon73.controller.domain.MessageResponse;
 import com.example.zvon73.entity.Bell;
 import com.example.zvon73.entity.BellTower;
 import com.example.zvon73.entity.Enums.BellStatus;
+import com.example.zvon73.entity.Temple;
+import com.example.zvon73.entity.User;
 import com.example.zvon73.repository.BellRepository;
 import com.example.zvon73.service.Exceptions.ValidateException;
 import lombok.RequiredArgsConstructor;
@@ -23,11 +25,17 @@ public class BellService {
 
     private final BellRepository bellRepository;
     private final BellTowerService bellTowerService;
+    private final UserService userService;
+
+
     @Transactional
     public Bell create(BellDto bell)
     {
-
+        User currentUser = userService.getCurrentUser();
         BellTower bellTower = bellTowerService.findById(UUID.fromString(bell.getBellTowerId()));
+
+        if(!bellTower.getTemple().checkRinger(currentUser.getId()))
+            throw new RuntimeException("403 : Not access");
 
         Bell newBell = Bell.builder()
                 .title(bell.getTitle())
@@ -45,52 +53,61 @@ public class BellService {
         return bellRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Колокола с данным id не найден"));
     }
-    private void validate(BellDto bell){
-        if(bell.getTitle().isEmpty() || bell.getTitle().length() < 5 || bell.getTitle().length() >100)
-            throw new ValidateException("Некорректное название колокола");
-        if(bell.getWeight() < 0)
-            throw new ValidateException("Некорректной вес колокола");
-        if(bell.getManufacturer().isEmpty() || bell.getManufacturer().length() < 3 || bell.getManufacturer().length() > 100)
-            throw new ValidateException("Некорректное название производителя");
-        if(bell.getImage().length == 0)
-            throw new ValidateException("Некорректное изображение колокола");
-        if(bell.getSound().length == 0)
-            throw new ValidateException("Некорректный звук колокола");
-        if(bell.getBellTowerId().toString().isEmpty())
-            throw new ValidateException("Некорректный номер колокольни");
-    }
+
 
     @Transactional
     public Bell update(BellDto newBell)
     {
+        User currentUser = userService.getCurrentUser();
         Bell currentBell = findById(UUID.fromString(newBell.getId()));
-            if( !currentBell.getTitle().equals(newBell.getTitle()) )
-                currentBell.setTitle(newBell.getTitle());
 
-            if( !currentBell.getManufacturer().equals(newBell.getManufacturer()) )
-                currentBell.setManufacturer(newBell.getManufacturer());
+        if(!currentBell.getBellTower().getTemple().checkRinger(currentUser.getId()))
+            throw new RuntimeException("403 : Not access");
 
-            if( currentBell.getWeight() != (newBell.getWeight()) )
-                currentBell.setWeight(newBell.getWeight());
+        if( !currentBell.getTitle().equals(newBell.getTitle()) )
+            currentBell.setTitle(newBell.getTitle());
 
-            if( currentBell.getBellTower().getId() != (UUID.fromString(newBell.getBellTowerId()))) {
-                BellTower newBellTower = bellTowerService.findById(UUID.fromString(newBell.getBellTowerId()));
-                currentBell.setBellTower(newBellTower);
+        if( !currentBell.getManufacturer().equals(newBell.getManufacturer()) )
+            currentBell.setManufacturer(newBell.getManufacturer());
+
+        if( currentBell.getWeight() != (newBell.getWeight()) )
+            currentBell.setWeight(newBell.getWeight());
+
+        if( currentBell.getBellTower().getId() != (UUID.fromString(newBell.getBellTowerId()))) {
+            BellTower newBellTower = bellTowerService.findById(UUID.fromString(newBell.getBellTowerId()));
+            currentBell.setBellTower(newBellTower);
             }
 
-            if ( !Arrays.equals(currentBell.getSound(), newBell.getSound()) )
-                currentBell.setSound(newBell.getSound());
-
-            if( !Arrays.equals(currentBell.getImage(), newBell.getImage()) )
-                currentBell.setImage(newBell.getImage());
 
         return bellRepository.save(currentBell);
     }
 
     @Transactional
+    public MessageResponse madeCanned(UUID id){
+        try {
+            User currentUser = userService.getCurrentUser();
+            Bell currentBell = findById(id);
+
+            if(!currentBell.getBellTower().getTemple().checkRinger(currentUser.getId()))
+                return new MessageResponse("", "403 : Not access");
+
+            currentBell.setCanned(true);
+            bellRepository.save(currentBell);
+            return new MessageResponse("Колокол успешно списан", "");
+        }catch (Exception e){
+            return new MessageResponse("", e.getMessage());
+        }
+    }
+
+    @Transactional
     public MessageResponse delete(UUID id){
         try {
+            User currentUser = userService.getCurrentUser();
             Bell currentBell = findById(id);
+
+            if(!currentBell.getBellTower().getTemple().checkRinger(currentUser.getId()))
+                return new MessageResponse("", "403 : Not access");
+
             bellRepository.delete(currentBell);
             return new MessageResponse("Колокол успешно удалён", "");
         }catch (Exception e){
@@ -100,7 +117,7 @@ public class BellService {
 
     @Transactional(readOnly = true)
     public List<BellDto> findAll(){
-        return bellRepository.findAll().stream().map(BellDto::new).collect(Collectors.toList());
+        return bellRepository.findByCannedFalse().stream().map(BellDto::new).collect(Collectors.toList());
     }
 
     @Transactional(readOnly = true)
